documents_strands_agent/requirements.txt
strands-agents>=0.1.0
strands-agents-builder>=0.1.0
bedrock-agentcore>=0.0.3
boto3>=1.35.0
